Only set items in admin context.
Quote has strange behavior where it might duplicate items if they are set in the quotes setItems, instead of in the item collection.
When creating an order in admin this isn't the case, instead a different bug appears where quote items is empty but items collection is not.

This will ensure the bugfix for the admin problem is only applied in the admin context.